requirements.txt
accelerate==1.3.0
deep_translator==1.11.4
diffusers==0.32.0
huggingface_hub==0.29.1
openai==1.63.0
Pillow==11.1.0
pydantic==2.10.6
python-dotenv==1.0.1
transformers==4.48.3
# torch==2.5.1+cu124
# torch==2.6.0
